+++ src/components/Shared/Header/HeaderMenu.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Link } from 'react-router-dom';
-import { logoutAction } from "../../../redux/actions";
+import { logoutAction } from "../../../modules/Auth/actions";
 import { useDispatch } from "react-redux";
 
 const HeaderMenu = () => {
