+++ src/modules/Auth/actions.js
@@ -1,0 +1,7 @@
+import {createAction} from "redux-actions";
+import {authFailure as failure, authRequest as auth, authSuccess as success, logoutAction as logout} from "../Shared/types";
+
+export const fetchAuthRequest = createAction(auth);
+export const fetchAuthSuccess = createAction(success);
+export const fetchAuthFailure = createAction(failure);
+export const logoutAction = createAction(logout);
