+++ src/modules/Map/index.js
@@ -1,0 +1,3 @@
+export {address, route} from "./reducers"
+export { default as sagas } from "./sagas"
+export * from "./actions"
