+++ src/modules/Profile/index.js
@@ -1,0 +1,3 @@
+export { profile } from "./reducers"
+export { default as sagas } from "./sagas"
+export * from "./actions"
