+++ src/modules/Profile/reducers.js
@@ -1,0 +1,7 @@
+import {handleActions} from "redux-actions";
+import {fetchProfileSuccess} from "./actions";
+
+export const profile = handleActions({
+    [fetchProfileSuccess]: (state, action) => action.payload
+}, null);
+
