+++ src/modules/index.js
@@ -1,0 +1,24 @@
+import { combineReducers } from "redux";
+import { fork } from "redux-saga/effects";
+import { isLoggedIn } from './Auth';
+import { profile } from './Profile';
+import { address, route } from './Map';
+import { pending, error} from './Shared/reducers';
+import { sagas as authSaga } from "./Auth/";
+import { sagas as profileSaga } from "./Profile/";
+import { sagas as MapSaga } from "./Profile/";
+
+export default combineReducers({
+    isLoggedIn,
+    profile,
+    address,
+    route,
+    pending,
+    error
+});
+
+export function* rootSaga () {
+    yield fork(authSaga);
+    yield fork(profileSaga);
+    yield fork(MapSaga);
+}
